Fix a bug causing the installation to fail
By default docker (and docker compose) only wait for the container to
boot, not for the underlying service to be up. If the install script
moves on fast enough, it will try to connect to psql before the psql
listens on his port, causing a crash.

A simple solution would have been to add a small sleep in the makefile,
as psql is fast to boot (unlike other databases) and the issue will not
trigger often.

A better solution is to add a loop that waits for the service to be up.

If using a node container instead of the host's node, one could use the
healthcheck and status condition features to have the node container
wait for the db to be available before booting. But this is not the case
here.